+++ syscalls/missiles/mp1-taux.h
@@ -1,8 +1,6 @@
 #ifndef _MP1_TAUX_H
 #define _MP1_TAUX_H
 
-#include <stdint.h>
-
 #define TUX_SET_LED     0x10
 #define TUX_BUTTONS     0x12
 #define TUX_INIT        0x13
@@ -20,10 +18,10 @@ enum {
     TB_ALL    = 0xff,
 };
 
-void taux_display_str(int32_t taux_fd, const char *str);
-void taux_display_time(int32_t taux_fd, int32_t num_seconds);
-void taux_display_coords(int32_t taux_fd, int32_t x, int32_t y);
-void taux_display_num(int32_t taux_fd, int32_t score);
-uint8_t taux_get_input(int32_t taux_fd);
+void taux_display_str(int taux_fd, const char *str);
+void taux_display_time(int taux_fd, int num_seconds);
+void taux_display_coords(int taux_fd, int x, int y);
+void taux_display_num(int taux_fd, int score);
+int taux_get_input(int taux_fd);
 
 #endif /* _MP1_TAUX_H */
